refactor(api): share implementation for http
