+++ vms/saevm/cchain/http.go
@@ -5,12 +5,10 @@ package cchain
 
 import (
 	"context"
-	"errors"
 	"fmt"
 	"net/http"
 
 	"github.com/ava-labs/avalanchego/graft/evm/utils/rpc"
-	"github.com/ava-labs/avalanchego/utils"
 	"github.com/ava-labs/avalanchego/vms/saevm/sae"
 )
 
@@ -25,7 +23,7 @@ var handlerPaths = append(sae.HandlerPaths, avaxHTTPExtensionPath)
 // augmented with the avax service. None of the handlers are usable until after
 // the [VM] is set as bootstrapping/normal operation.
 func (vm *VM) CreateHandlers(context.Context) (map[string]http.Handler, error) {
-	return vm.handlers.toInterface(), nil
+	return vm.handlers.AsInterface(), nil
 }
 
 // setHandlers initializes the lazy handlers with the real implementations.
@@ -46,75 +44,6 @@ func (vm *VM) setHandlers(ctx context.Context) error {
 
 	m[avaxHTTPExtensionPath] = handler
 
-	return vm.handlers.setHandlers(m)
-}
-
-// A handlerMap is a fixed set of HTTP routes whose implementations can be
-// supplied after the routes have been exposed.
-type handlerMap map[string]*lazyHandler
-
-func newHandlerMap(paths ...string) handlerMap {
-	m := make(handlerMap, len(paths))
-	for _, path := range paths {
-		m[path] = &lazyHandler{}
-	}
-	return m
-}
-
-func (m handlerMap) toInterface() map[string]http.Handler {
-	iface := make(map[string]http.Handler, len(m))
-	for path, lazy := range m {
-		iface[path] = lazy
-	}
-	return iface
-}
-
-var (
-	errUnregisteredHandlerPath = errors.New("handler path not registered at construction")
-	errMissingHandlerPath      = errors.New("no handler provided for registered path")
-)
-
-// setHandlers routes each lazy handler to its actual implementation. The
-// actual paths MUST exactly match the paths registered at construction.
-func (m handlerMap) setHandlers(actual map[string]http.Handler) error {
-	for path := range actual {
-		if _, ok := m[path]; !ok {
-			return fmt.Errorf("%w: %q", errUnregisteredHandlerPath, path)
-		}
-	}
-	for path := range m {
-		if _, ok := actual[path]; !ok {
-			return fmt.Errorf("%w: %q", errMissingHandlerPath, path)
-		}
-	}
-
-	for path, h := range actual {
-		if lazy, ok := m[path]; ok {
-			lazy.set(h)
-		}
-	}
+	vm.handlers.Set(m)
 	return nil
-}
-
-var _ http.Handler = (*lazyHandler)(nil)
-
-// A lazyHandler is a placeholder for an actual HTTP handler. It returns 404
-// until the actual handler is set.
-type lazyHandler struct {
-	h utils.Atomic[http.Handler]
-}
-
-func (l *lazyHandler) ServeHTTP(w http.ResponseWriter, r *http.Request) {
-	h := l.h.Get()
-
-	if h == nil {
-		http.NotFound(w, r)
-		return
-	}
-
-	h.ServeHTTP(w, r)
-}
-
-func (l *lazyHandler) set(h http.Handler) {
-	l.h.Set(h)
 }
